Read OPDS feed & Add simple shared state management
NB: A full-scale Flux implementation seemed like overkill for what is
essentially a toy at the moment. So in order to handle shared state I'm
following the practices outlined here:
https://vuejs.org/v2/guide/state-management.html#Simple-State-Management-from-Scratch

This begins an experiment in consuming OPDS data to power the library that this
app provides. This is an XML format instead of a JSON api, but Vue can handle
this data fine with a little help parsing the XML.

When the app loads for the first time, the root Vue instance calls a function
defined in the state.js file: `fetchBooksAction(url)`. This sends a request to
the URL expecting an XML response. The xml2js module is used to parse the XML
data. Note the parser options in this function: setting these values makes the
parsed response behave a lot more like what you would get from a JSON
api (otherwise everything is wrapped in an extraneous array, and there are lots
of weird properties for unimportent attributes on the XML tags).

The key is that this resulting data, once parsed, is stored in a central
location in the `store` object. Any component which needs to access (read-only
please!) the data in this store should import this file.

So in any component that needs to access the complete books data, it should have
a data property that looks like this:

```
data () {
  return {
    books: store.state.books
    // other local state can live here as well
    ...
  }
}
```

For the time being the books array is the only thing that should need to be
shared among multiple components.